// src/app/_dto/TblmDocAttach.dto.ts
export class    TblmDocAttach {
    HmDoc_ID: number;
    HmDoc_SysID: number = 0;
    HmDoc_Tab_ID: string = null;
    HmDoc_Code: string = null;
    HmDoc_Name: string = null;
    HmDoc_Path: string = null;
    HmDoc_Start_Date: Date = null;
    HmDoc_Expiry_Date: Date = null;
    HmDoc_Alert_Date: Date = null;
    HmDoc_Emp_SysID: number = 0;
    HmDoc_Emp_Code: string = null;
    HmDoc_Emp_Name: string = null;
    HmDoc_Emp_Email: string = null;
    HmDoc_AlertRep_Hrs: string = null;
    HmDoc_Alert_SysID: number = 0;
    HmDoc_Alert_Code: string = null;
    HmDoc_Alert_Name: string = null;
    HmDoc_Alert_Stop_Yn: string = null;
    HmDoc_GridNo: number = null;    
   





}